backend/requirements.txt
openai>=1.0.0
pymongo>=4.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
pydantic>=2.10.0
